requirements.txt
APScheduler==3.10.1
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.5.8
Django==5.0.2
idna==3.7
marshmallow==3.21.2
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
packaging==24.0
phonepe-sdk==1.1.0
pycryptodome==3.18.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==1.26.12
